Feat Tags API (#2)
* Add tagging functionality to items and stores

This commit introduces tagging functionality to the application by introducing a relationship between tags and items and between tags and stores.

* Add tag functionality to store and item models

Added a new module for handling tag operations: creating, deleting, and linking tags to items in stores. Introduced new fields in the StoreSchema and ItemSchema to accommodate the tags.