Add JWT blacklist and update main app settings
Added a new `blacklist.py` file in the `jwt` app directory to manage the blacklist of the JWT tokens. This blacklist will be used to add tokens when a user logs out. Updated `main.py` to import this new blacklist. Extended Flask main app settings to include JWT authentication configurations. Importantly, token checkers to assess if tokens are in blacklist are included. Additional error handling for a variety of token issues has been added for better user management and security.